Crystalin split executable (#105)
* Adds name as parameter for collator nodes

* Split index into executable and logic file

* Adds pre-install to the package.json

* Replace preinstall by prepare

* Split index and cli file

* Removes types from index

* Adds types to the package.json

* Adds compiled types

* lint

* Update package.json
