+++ src/rpc.ts
@@ -19,7 +19,11 @@ filterConsole([
 // TODO: Add a timeout where we know something went wrong so we don't wait forever.
 export async function connect(port: number, types: any) {
 	const provider = new WsProvider("ws://127.0.0.1:" + port);
-	const api = await ApiPromise.create({ provider, types, throwOnConnect: false });
+	const api = await ApiPromise.create({
+		provider,
+		types,
+		throwOnConnect: false,
+	});
 	return api;
 }
